login

+++ studybud/basee/views.py
@@ -1,12 +1,17 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
 from django.contrib.auth.models import User
-from django.contrib.auth import authenticate, login
+from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Room, Topic
 from .forms import RoomForm
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -27,6 +32,10 @@ def loginPage(request):
     context = {}
     return render(request, 'basee/login_register.html', context)
 
+def logoutUser(request):
+    logout(request)
+    return redirect('login')
+
 def home(request):
     action = request.GET.get('action') if request.GET.get('action') != None else ''
     
@@ -46,6 +55,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'basee/room.html', context)
 
+@login_required(login_url ='login' )
 def createRoom(request):
     form = RoomForm()
     
@@ -58,9 +68,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'basee/room_form.html',context)
 
+@login_required(login_url ='login' )
 def updateRoom(request, pk):
     room = Room.objects.get(id = pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed to do that !!')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -72,8 +86,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'basee/room_form.html', context)
 
+@login_required(login_url ='login' )
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed to do that !!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
